refactor(allocations): keep route policy separate from read mechanics
Move cursor semantics and allocation read assembly behind focused services so the router retains validation, authorization, response mapping, and unchanged write behavior.

Constraint: Preserve allocation paths, cursor order, fallback behavior, optimistic locking, and response envelopes without dependency or persistence changes.

Rejected: Extract allocation writes with reads | write semantics are explicitly outside this pure extraction lane.

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep cursor codec and SQL predicate changes synchronized with direct characterization tests.

Tested: 103 focused server tests; TypeScript baseline check; full lint and guardrails; production build; git diff check.

Not-tested: API suite remains feature-gated; DB cursor integration blocked by stale local schema; full unit command timed out after 10 minutes.
